Potential fix for code scanning alert no. 29: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using double precision to avoid overflow. This can be achieved by casting the float values to double before performing the multiplication. This way, the multiplication will be done using double precision, and the result will not overflow.

  - Cast the float values ToLinearF[j] and ToLinearF[j+1] to double before performing the multiplication.
  - Update the code on line 558 to use the casted values.

+++ libtiff-v4.0.10/libtiff/tif_pixarlog.c
@@ -555,7 +555,7 @@ PixarLogMakeTables(PixarLogState *sp)
 
     j = 0;
     for (i = 0; i < lt2size; i++)  {
-	if ((i*linstep)*(i*linstep) > ToLinearF[j]*ToLinearF[j+1])
+	if ((i*linstep)*(i*linstep) > (double)ToLinearF[j]*(double)ToLinearF[j+1])
 	    j++;
 	FromLT2[i] = (uint16)j;
     }
